requirements.txt
django
djangorestframework
djongo
django-cors-headers
dnspython
django-environ
pymongo==3.12.3


